+++ __tests__/html/accessibility.liveRegionAttachment.adaptiveCard.js
@@ -1,0 +1,9 @@
+/**
+ * @jest-environment ./__tests__/html/__jest__/WebChatEnvironment.js
+ */
+
+describe('accessibility requirement', () => {
+  describe('attachments in live region', () => {
+    test('Adaptive Card', () => runHTMLTest('accessibility.liveRegionAttachment.adaptiveCard.html'));
+  });
+});