+++ packages/component/src/Middleware/AttachmentForScreenReader/VideoAttachment.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import useLocalizer from '../../hooks/useLocalizer';
+
+const VideoAttachment = () => {
+  const localize = useLocalizer();
+
+  const label = localize('ATTACHMENT_VIDEO');
+
+  return <div>{label}</div>;
+};
+
+export default VideoAttachment;